+++ app/controllers/jobs_controller.rb
@@ -1,7 +1,7 @@
 class JobsController < ApplicationController
 
   def index
-    @jobs = Job.all_open
+    @jobs = Job.vacant
   end
 
   def show
@@ -9,6 +9,7 @@ class JobsController < ApplicationController
   end
 
   def create
+    redirect_to musos_path and return unless !!current_host
     @job = Job.new(job_create_params)
     if @job.save
       redirect_to '/jobs'
@@ -18,11 +19,13 @@ class JobsController < ApplicationController
   end
 
   def edit
+    redirect_to musos_path and return unless !!current_host
     @job = Job.find(params[:id])
   end
 
   def update
-    @job = Job.find(params[:id])
+    @job = Job.find(params[:id])''
+    redirect_to job_path(job.id) and return unless my_job?(@job)
     if @job.update_attributes(job_edit_params)
       redirect_to '/jobs'
     else
@@ -32,6 +35,7 @@ class JobsController < ApplicationController
 
   def destroy
     @job = Job.find(params[:id])
+    redirect_to job_path(job.id) and return unless !job.complete && my_job?(@job)
     if @job.destroy
       redirect_to "/jobs"
     else
@@ -41,6 +45,7 @@ class JobsController < ApplicationController
 
   def select_successful_applicant
     job = Job.find(params[:id])
+    redirect_to job_path(job.id) and return unless job.filled && my_job?(@job)
     job.job_application = JobApplication.find(params[:job_application_id])
     job.filled = true
     if job.save
@@ -54,6 +59,7 @@ class JobsController < ApplicationController
 
   def mark_job_completed
     job = Job.find(params[:id])
+    redirect_to job_path(job.id) and return unless job.filled && my_job?(@job)
     job.complete = true
      if job.save
       NotificationMailer.completed_job_app_email(job).deliver_later
